requirements.txt
torch~=1.12.1
torchvision~=0.13.1
numpy~=1.22.3
pandas~=1.4.2
pillow~=9.1.1
opencv-python~=4.6.0
openslide-python~=1.3.1
wget~=3.2
scikit-learn~=1.1.2